# a7_openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.0
  title: 'LBAW AltArt Web API'
  description: 'Web Resources Specification (A7) for AltArt'
servers:
  - url: http://lbaw2123.lbaw-prod.fe.up.pt/
    description: Production server
...
paths:
...
  /user/{id}:
    get:
      operationId: R106
      summary: 'R106: View User Profile'
      description: 'Provide user profile. Access: USR'
      tags:
        - 'M01: Authentication and Individual Profile'
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '200':
          description: 'Ok. Show [UI06](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2123/-/wikis/er#ui06-users-profile)'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    username:
                      type: string
                    bio:
                      type: string
                    instagram:
                      type: string
                    twitter:
                      type: string
                    linkedin:
                      type: string
                    facebook:
                      type: string
                    profile_photo:
                      type: string
                    number_followers:
                      type: integer
                    number_following:
                      type: integer
                    number_likes:
                      type: integer
                    posts:
                      type: array
                      items:
                        type: string
                      nullable: true
                  example:
                    - name: "Ana Sousa"
                      username: "anasousa"
                      bio: "Hi!"
                      instagram: https://www.instagram.com/
                      twitter: https://www.twitter.com/
                      linkedin: https://www.linkedin.com/
                      facebook: https://www.facebook.com/
                      profile_photo: null
                      number_followers: 0
                      number_following: 20
                      number_likes: 0
                      posts: null

        '404':
          description: 'Page not found: Show page not found UI'

    delete:
      operationId: R107
      summary: "R107: Delete account action"
      description: "Delete account. Access: OWN"
      tags:
        - 'M01: Authentication and Individual Profile'
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '302':
          description: 'Redirect after deleting user.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful account deletion. Redirect to homepage.'
                  value: '/'
        '403':
          description: 'Forbidden client: show no permission UI.'
...